examples/realtime-next/server/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
